pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fundus-evaluation"
version = "1.0.0"
authors = [
    { name = "Max Dallabetta", email = "max.dallabetta@googlemail.com" },
    { name = "Conrad Dobberstein" },
    { name = "Adrian Breiding" },
    { name = "Alan Akbik" },
]
description = "Evaluation of the Fundus news crawler"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
]
keywords = ["web scraping, web crawling", "evaluation"]
dependencies = [
    "more-itertools==9.1.0",
    "pandas==2.0.3",
    "tqdm==4.66.2",
    "typing-extensions==4.10.0",
    # News crawler dependencies
    "boilerpipe3==1.3.0",
    "fundus@git+https://github.com/flairNLP/fundus.git@b276878207bd6bf89d92015340434e6e3d9bb71a",
    "justext==3.0.0",
    "news-please==1.5.44",
    "trafilatura==1.7.0",
    # Scorer dependencies
    "jiwer==3.0.3",
    "rouge-score==0.1.2",
]

[project.urls]
"Repository" = "https://github.com/dobbersc/fundus-evaluation"
"Fundus Repository" = "https://github.com/flairNLP/fundus"

[project.optional-dependencies]
dev = [
    "mypy==1.8.0",
    "isort==5.13.2",
    "black==24.2.0",
    # Type stubs
    "pandas-stubs==2.0.3.230814",
    "types-tqdm==4.66.0.20240106",
]

[project.scripts]
evaluate = "fundus_evaluation:__main__.main"

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
no_implicit_optional = true
no_implicit_reexport = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "boilerpipe.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fundus.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "fundus_evaluation.scrapers.bte"
ignore_errors = true

[[tool.mypy.overrides]]
module = "jiwer.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "justext.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "newsplease.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "nltk.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rouge_score.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "trafilatura.*"
ignore_missing_imports = true

[tool.black]
line-length = 120
target-version = ['py38']

[tool.isort]
profile = "black"
